Added date param trait
