+++ includes/core/server/request/class-list-transactions.php
@@ -19,7 +19,7 @@ use WP_REST_Request;
  */
 class List_Transactions extends Paginated {
 
-	use Order_Info;
+	use Date_Parameters, Order_Info;
 
 	const DEFAULT_PARAMS = [
 		'sort'      => 'date',
@@ -151,47 +151,6 @@ class List_Transactions extends Paginated {
 	public function set_search( string $search ) {
 		$this->set_param( 'search', $search );
 	}
-
-	/**
-	 * Set date after.
-	 *
-	 * @param string $date_after Date after.
-	 *
-	 * @return void
-	 * @throws \WCPay\Core\Exceptions\Server\Request\Invalid_Request_Parameter_Exception
-	 */
-	public function set_date_after( string $date_after ) {
-		$this->validate_date( $date_after );
-		$this->set_param( 'date_after', $date_after );
-	}
-
-	/**
-	 * Set date before.
-	 *
-	 * @param string $date_before Date before.
-	 *
-	 * @return void
-	 * @throws \WCPay\Core\Exceptions\Server\Request\Invalid_Request_Parameter_Exception
-	 */
-	public function set_date_before( string $date_before ) {
-		$this->validate_date( $date_before );
-		$this->set_param( 'date_before', $date_before );
-	}
-
-	/**
-	 * Set date between.
-	 *
-	 * @param array $date_between Date between.
-	 *
-	 * @return void
-	 */
-	public function set_date_between( array $date_between ) {
-		// Add only if it's not empty.
-		if ( $date_between ) {
-			$this->set_param( 'date_between', $date_between );
-		}
-	}
-
 
 	/**
 	 * Set type is.
